Update linuxGPUDetector.ts
Change regex for "clinfo" to be able to detect GPUs in a system with both AMD and Nvidia GPUs.

Explanation: Instead of just grabbing the group as a single number, grab the AMD/NV part in each line.
This is required since on a system with both AMD and NVIDIA GPUs, GPUs with the same index will be overwritten. Using AMD/NV as part of the key prevents that.